multimodal-rag-pipeline/requirements.txt
boto3
botocore
tabula-py
faiss-cpu
pymupdf
requests
langchain-aws
langchain-text-splitters
numpy
tqdm
PyYAML
ipython
jupyter
